Feat: 다른 용량 단위에 대한 구매이력 추가 후 구매 예정일 계산
- 실패하는 테스트 케이스 작성

+++ src/main/java/timo/shoppingcycle/domain/PurchaseHistory.java
@@ -5,18 +5,18 @@ import java.time.LocalDate;
 public class PurchaseHistory {
 
     private final String name;
-    private final int capacity;
+    private final double capacity;
     private Unit unit;
 
     private final LocalDate purchaseDate;
 
-    public PurchaseHistory(String name, int capacity, LocalDate purchaseDate) {
+    public PurchaseHistory(String name, double capacity, LocalDate purchaseDate) {
         this.name = name;
         this.capacity = capacity;
         this.purchaseDate = purchaseDate;
     }
 
-    public PurchaseHistory(String name, int capacity, Unit unit, LocalDate purchaseDate) {
+    public PurchaseHistory(String name, double capacity, Unit unit, LocalDate purchaseDate) {
         this.name = name;
         this.capacity = capacity;
         this.unit = unit;
@@ -27,7 +27,7 @@ public class PurchaseHistory {
         return purchaseDate;
     }
 
-    public int getCapacity() {
+    public double getCapacity() {
         return capacity;
     }
